# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RequestKind = typeof RequestKind[keyof typeof RequestKind];


export const RequestKind = {
  contact: 'contact',
  order: 'order',
} as const;

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];


export const RequestStatus = {
  new: 'new',
  in_progress: 'in_progress',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Request {
  id: number;
  kind: RequestKind;
  name: string;
  contact: string;
  message: string;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  budget?: string | null;
  status: RequestStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type RequestInputKind = typeof RequestInputKind[keyof typeof RequestInputKind];


export const RequestInputKind = {
  contact: 'contact',
  order: 'order',
} as const;

export interface RequestInput {
  kind: RequestInputKind;
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  contact: string;
  /** @minLength 5 */
  message: string;
  service?: string;
  budget?: string;
}

export type RequestUpdateStatus = typeof RequestUpdateStatus[keyof typeof RequestUpdateStatus];


export const RequestUpdateStatus = {
  new: 'new',
  in_progress: 'in_progress',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface RequestUpdate {
  status?: RequestUpdateStatus;
  notes?: string;
}

export interface Dashboard {
  total: number;
  newCount: number;
  orderCount: number;
  completedCount: number;
  recent: Request[];
}

export interface TelegramLogin {
  id: number;
  first_name: string;
  last_name?: string;
  username?: string;
  photo_url?: string;
  auth_date: number;
  hash: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  adminId: number | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  username?: string | null;
}

export interface TelegramBot {
  username: string;
}

export type ListRequestsParams = {
status?: ListRequestsStatus;
kind?: ListRequestsKind;
search?: string;
};

export type ListRequestsStatus = typeof ListRequestsStatus[keyof typeof ListRequestsStatus];


export const ListRequestsStatus = {
  new: 'new',
  in_progress: 'in_progress',
  completed: 'completed',
  archived: 'archived',
} as const;

export type ListRequestsKind = typeof ListRequestsKind[keyof typeof ListRequestsKind];


export const ListRequestsKind = {
  contact: 'contact',
  order: 'order',
} as const;