# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Request } from './request';

export interface Dashboard {
  total: number;
  newCount: number;
  orderCount: number;
  completedCount: number;
  recent: Request[];
}