# lib/api-zod/src/generated/types/request.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestKind } from './requestKind';
import type { RequestStatus } from './requestStatus';

export interface Request {
  id: number;
  kind: RequestKind;
  name: string;
  contact: string;
  message: string;
  /** @nullable */
  service?: string | null;
  /** @nullable */
  budget?: string | null;
  status: RequestStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}