# lib/api-zod/src/generated/types/requestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestInputKind } from './requestInputKind';

export interface RequestInput {
  kind: RequestInputKind;
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  contact: string;
  /** @minLength 5 */
  message: string;
  service?: string;
  budget?: string;
}