# lib/api-zod/src/generated/types/requestUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RequestUpdateStatus } from './requestUpdateStatus';

export interface RequestUpdate {
  status?: RequestUpdateStatus;
  notes?: string;
}